moved constants and enums out of utils
