+++ src/UXClient/Components/Grid/Grid.ts
@@ -1,6 +1,7 @@
 import * as d3 from 'd3';
 import './Grid.scss';
-import Utils, { DataTypes } from "./../../Utils";
+import Utils from "./../../Utils";
+import { DataTypes } from "./../../Constants/Enums";
 import { Component } from "./../../Interfaces/Component";
 import { ChartOptions } from '../../Models/ChartOptions';
 import { ChartComponentData } from '../../Models/ChartComponentData';
