+++ src/UXClient/Components/Marker/Marker.ts
@@ -1,12 +1,13 @@
 import * as d3 from 'd3';
 import './Marker.scss';
-import Utils, { LINECHARTCHARTMARGINS, DataTypes, MARKERVALUENUMERICHEIGHT, LINECHARTXOFFSET, TooltipMeasureFormat} from "./../../Utils";
+import Utils from "./../../Utils";
+import { LINECHARTCHARTMARGINS, MARKERVALUENUMERICHEIGHT, LINECHARTXOFFSET } from "./../../Constants/Constants";
 import {Component} from "./../../Interfaces/Component";
 import { ChartOptions } from '../../Models/ChartOptions';
 import { LineChartData } from '../../Models/LineChartData';
 import { Tooltip } from '../Tooltip/Tooltip';
 import { ChartComponentData } from '../../Models/ChartComponentData';
-import { KeyCodes, ShiftTypes } from '../../Constants/Enums';
+import { KeyCodes, ShiftTypes, DataTypes, TooltipMeasureFormat } from '../../Constants/Enums';
 
 const MARKERSTRINGMAXLENGTH = 250;
 const MARKERVALUEMAXWIDTH = 80;
