+++ src/UXClient/Components/PieChart/PieChart.ts
@@ -1,6 +1,7 @@
 import * as d3 from 'd3';
 import './PieChart.scss';
-import Utils, {TooltipMeasureFormat} from "./../../Utils";
+import Utils from "./../../Utils";
+import { TooltipMeasureFormat } from "./../../Constants/Enums";
 import {Legend} from './../Legend/Legend';
 import {ContextMenu} from './../ContextMenu/ContextMenu';
 import { ChartComponentData } from '../../Models/ChartComponentData';
