+++ src/UXClient/Constants/Constants.ts
@@ -24,3 +24,17 @@ export const swimlaneLabelConstants = {
 }
 
 export const CharactersToEscapeForExactSearchInstance = ['"', '`', '\'', '!', '(', ')', '^', '[', '{', ':', ']', '}', '~', '/', '\\', '@', '#', '$', '%', '&', '*', ';', '=', '.', '_', '-', '<', '>', ',', '?'];
+
+export const NONNUMERICTOPMARGIN = 8;
+export const LINECHARTTOPPADDING = 16;
+export const GRIDCONTAINERCLASS = 'tsi-gridContainer';
+export const LINECHARTCHARTMARGINS = {
+    top: 40,
+    bottom: 40,
+    left: 70,
+    right: 60
+};
+export const LINECHARTXOFFSET = 8;
+export const MARKERVALUENUMERICHEIGHT = 20;
+export const VALUEBARHEIGHT = 3;
+export const SERIESLABELWIDTH = 92;
