+++ src/UXClient/Interfaces/Plot.ts
@@ -1,5 +1,6 @@
 import {Component} from "./Component";
-import Utils, { NONNUMERICTOPMARGIN } from '../Utils';
+import Utils from '../Utils';
+import { NONNUMERICTOPMARGIN } from "../Constants/Constants";
 
 class Plot extends Component {
     protected chartHeight;
